refactor: consolidate ObjectIntBiConsumer into common-lang util.function
binding-http and engine each carried their own copy of this helper
functional interface. Helper functional interfaces of this shape
already live in common-lang's util.function package (LongObjectBiFunction,
LongObjectPredicate), so ObjectIntBiConsumer moves there too and both
call sites reference the shared type instead of a local duplicate.

engine now depends on common-lang (a dependency-free leaf module, no
cycle risk) and re-exports it transitively so any Router implementation
can reference ObjectIntBiConsumer; binding-http picks it up the same
way it already does common-agrona, transitively through its existing
engine dependency, with no pom.xml or module-info change needed there.
