--- runtime/binding-http/src/main/java/io/aklivity/zilla/runtime/binding/http/internal/util/function/ObjectIntBiConsumer.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2021-2026 Aklivity Inc.
- *
- * Aklivity licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package io.aklivity.zilla.runtime.binding.http.internal.util.function;
-
-import java.util.Objects;
-import java.util.function.BiConsumer;
-
-@FunctionalInterface
-public interface ObjectIntBiConsumer<T> extends BiConsumer<T, Integer>
-{
-    void accept(T t, int value);
-
-    @Override
-    default void accept(T t, Integer value)
-    {
-        this.accept(t, value.intValue());
-    }
-
-    default ObjectIntBiConsumer<T> andThen(
-        ObjectIntBiConsumer<? super T> after)
-    {
-        Objects.requireNonNull(after);
-
-        return (r, l) ->
-        {
-            accept(r, l);
-            after.accept(r, l);
-        };
-    }
-}
