--- runtime/binding-http/src/test/java/io/aklivity/zilla/runtime/binding/http/internal/util/function/ObjectIntBiConsumerTest.java
@@ -1,56 +1,0 @@
-/*
- * Copyright 2021-2026 Aklivity Inc.
- *
- * Aklivity licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package io.aklivity.zilla.runtime.binding.http.internal.util.function;
-
-import static org.mockito.Mockito.spy;
-import static org.mockito.Mockito.verify;
-
-import java.util.Random;
-
-import org.junit.Test;
-
-public class ObjectIntBiConsumerTest
-{
-    @Test
-    @SuppressWarnings("unchecked")
-    public void shouldInvokePrimitiveAccept()
-    {
-        final ObjectIntBiConsumer<Object> consumer = spy(ObjectIntBiConsumer.class);
-
-        final Object object = new Object();
-        final Integer value = new Random().nextInt();
-
-        consumer.accept(object, value);
-
-        verify(consumer).accept(object, value.intValue());
-    }
-
-    @Test
-    @SuppressWarnings("unchecked")
-    public void shouldInvokeBeforeAndThenAfter()
-    {
-        final ObjectIntBiConsumer<Object> before = spy(ObjectIntBiConsumer.class);
-        final ObjectIntBiConsumer<Object> after = spy(ObjectIntBiConsumer.class);
-
-        final Object object = new Object();
-        final int value = new Random().nextInt();
-
-        before.andThen(after).accept(object, value);
-
-        verify(before).accept(object, value);
-        verify(after).accept(object, value);
-    }
-}
