+++ runtime/common-lang/src/main/java/io/aklivity/zilla/runtime/common/lang/util/function/ObjectIntBiConsumer.java
@@ -1,19 +1,18 @@
 /*
- * Copyright 2021-2026 Aklivity Inc.
+ * Copyright 2021-2026 Aklivity Inc
  *
- * Aklivity licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
+ * Licensed under the Aklivity Community License (the "License"); you may not use
+ * this file except in compliance with the License.  You may obtain a copy of the
+ * License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *   https://www.aklivity.io/aklivity-community-license/
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
  */
-package io.aklivity.zilla.runtime.engine.util.function;
+package io.aklivity.zilla.runtime.common.lang.util.function;
 
 import java.util.Objects;
 import java.util.function.BiConsumer;
